newer gitea doesn't always return owner as the first team
We cannot presume that owner will be the first team listed
so we need to run through all of the teams
After a lot of debugging, the old presumption that the Owner team,
will be the first team and therefor we can skip the first
returned team is not valid.
As a consequence if you have a LDAP group with the name Owner,
you will need to be carefull.

Signed-off-by: Toshaan Bharvani <toshaan@vantosh.com>